+++ packages/react/src/keys.ts
@@ -1,4 +1,9 @@
 import { Locale } from '@psitta/core';
 import { createContext } from 'react';
 
-export const LocaleContext = createContext<Locale | null>(null);
+type LocaleContext = [
+  Locale,
+  (locale: Locale) => void
+]
+
+export const LocaleContext = createContext<LocaleContext>([null as unknown as Locale, () => {}]);